# backend/src/routes/auth.routes.js
const express = require('express');
const bcrypt = require('bcrypt');
const jwt = require('jsonwebtoken');
const { mysqlPool } = require('../config/database');
const { authMiddleware } = require('../middleware/auth.middleware');

const router = express.Router();

const MAX_FIELD_LENGTH = {
  username: 50,
  title: 100,
  organization: 150,
  bio: 500
};

const sanitizeField = (value, limit) => {
  if (value === undefined || value === null) {
    return null;
  }
  const trimmed = String(value).trim();
  if (!trimmed) {
    return null;
  }
  return trimmed.length > limit ? trimmed.slice(0, limit) : trimmed;
};

/**
 * @swagger
 * /api/auth/register:
 *   post:
 *     summary: 注册新用户
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - username
 *               - email
 *               - password
 *             properties:
 *               username:
 *                 type: string
 *               email:
 *                 type: string
 *               password:
 *                 type: string
 *     responses:
 *       201:
 *         description: 用户创建成功
 *       400:
 *         description: 请求参数错误
 *       409:
 *         description: 用户名或邮箱已存在
 */
router.post('/register', async (req, res) => {
  try {
    const { username, email, password, organization, title, bio } = req.body;
    
    // 验证参数
    if (!username || !email || !password) {
      return res.status(400).json({ message: '用户名、邮箱和密码为必填项' });
    }
    
    // 检查用户名或邮箱是否已存在
    const [existingUsers] = await mysqlPool.execute(
      'SELECT * FROM users WHERE username = ? OR email = ?',
      [username, email]
    );
    
    if (existingUsers.length > 0) {
      return res.status(409).json({ message: '用户名或邮箱已存在' });
    }
    
    // 加密密码
    const salt = await bcrypt.genSalt(10);
    const hashedPassword = await bcrypt.hash(password, salt);
    
    const sanitizedOrganization = sanitizeField(organization, MAX_FIELD_LENGTH.organization);
    const sanitizedTitle = sanitizeField(title, MAX_FIELD_LENGTH.title);
    const sanitizedBio = sanitizeField(bio, MAX_FIELD_LENGTH.bio);

    // 创建新用户
    const [result] = await mysqlPool.execute(
      'INSERT INTO users (username, email, password_hash, role, organization, title, bio, created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
      [username, email, hashedPassword, 'user', sanitizedOrganization, sanitizedTitle, sanitizedBio, new Date()]
    );
    
    res.status(201).json({ 
      message: '用户创建成功',
      userId: result.insertId
    });
  } catch (error) {
    console.error('注册错误:', error);
    res.status(500).json({ message: '服务器内部错误' });
  }
});

/**
 * @swagger
 * /api/auth/login:
 *   post:
 *     summary: 用户登录
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - username
 *               - password
 *             properties:
 *               username:
 *                 type: string
 *                 description: 用户名或邮箱
 *               password:
 *                 type: string
 *     responses:
 *       200:
 *         description: 登录成功
 *       400:
 *         description: 请求参数错误
 *       401:
 *         description: 用户名或密码不正确
 */
router.post('/login', async (req, res) => {
  try {
    const { username, password } = req.body;
    
    // 验证参数
    if (!username || !password) {
      return res.status(400).json({ message: '用户名和密码为必填项' });
    }
    
    // 获取用户（允许用户名或邮箱登录）
    const [users] = await mysqlPool.execute(
      'SELECT * FROM users WHERE username = ? OR email = ?',
      [username, username]
    );
    
    if (users.length === 0) {
      return res.status(401).json({ message: '用户名或密码不正确' });
    }
    
    const user = users[0];

    // 验证密码
    const isPasswordValid = await bcrypt.compare(password, user.password_hash);
    
    if (!isPasswordValid) {
      return res.status(401).json({ message: '用户名或密码不正确' });
    }
    
    // 创建JWT令牌
    const token = jwt.sign(
      { 
        id: user.id, 
        username: user.username,
        email: user.email,
        role: user.role
      },
      process.env.JWT_SECRET,
      { expiresIn: '24h' }
    );
    
    // 记录登录日志
    await mysqlPool.execute(
      'INSERT INTO logs (user_id, action, target_id, timestamp) VALUES (?, ?, ?, ?)',
      [user.id, 'login', user.id, new Date()]
    );
    
    res.status(200).json({
      message: '登录成功',
      token,
      user: {
        id: user.id,
        username: user.username,
        email: user.email,
        role: user.role,
        organization: user.organization ?? null,
        title: user.title ?? null,
        bio: user.bio ?? null
      }
    });
  } catch (error) {
    console.error('登录错误:', error);
    res.status(500).json({ message: '服务器内部错误' });
  }
});

/**
 * @swagger
 * /api/auth/profile:
 *   get:
 *     summary: 获取当前用户信息
 *     tags: [Auth]
 *     security:
 *       - bearerAuth: []
 *     responses:
 *       200:
 *         description: 成功获取用户信息
 *       401:
 *         description: 未授权
 */
router.get('/profile', authMiddleware, async (req, res) => {
  try {
    if (!req.user || !req.user.id) {
      return res.status(401).json({ message: '未授权访问' });
    }

    const userId = req.user.id;

    const [users] = await mysqlPool.execute(
      'SELECT id, username, email, role, created_at, updated_at, organization, title, bio FROM users WHERE id = ?',
      [userId]
    );

    if (users.length === 0) {
      return res.status(404).json({ message: '未找到用户' });
    }

    const user = users[0];

    const [activityRows] = await mysqlPool.execute(
      `SELECT action, timestamp, details
       FROM logs
       WHERE user_id = ?
       ORDER BY timestamp DESC
       LIMIT 30`,
      [userId]
    );

    const lastLoginRecord = activityRows.find(row => row.action === 'login');

    const profile = {
      id: user.id,
      username: user.username,
      email: user.email,
      role: user.role,
      createdAt: user.created_at,
      updatedAt: user.updated_at,
      organization: user.organization ?? null,
      title: user.title ?? null,
      bio: user.bio ?? null,
      lastLogin: lastLoginRecord ? lastLoginRecord.timestamp : null,
      activities: activityRows.map(row => ({
        action: row.action,
        timestamp: row.timestamp,
        details: row.details || null
      }))
    };

    res.status(200).json(profile);
  } catch (error) {
    console.error('获取用户信息错误:', error);
    res.status(500).json({ message: '服务器内部错误' });
  }
});

/**
 * @swagger
 * /api/auth/profile:
 *   put:
 *     summary: 更新当前用户信息
 *     tags: [Auth]
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               username:
 *                 type: string
 *                 maxLength: 50
 *               organization:
 *                 type: string
 *                 maxLength: 150
 *               title:
 *                 type: string
 *                 maxLength: 100
 *               bio:
 *                 type: string
 *               password:
 *                 type: string
 *               confirmPassword:
 *                 type: string
 *     responses:
 *       200:
 *         description: 更新成功
 *       400:
 *         description: 请求参数错误
 *       401:
 *         description: 未授权
 */
router.put('/profile', authMiddleware, async (req, res) => {
  try {
    if (!req.user || !req.user.id) {
      return res.status(401).json({ message: '未授权访问' });
    }

    const userId = req.user.id;

    const {
      username,
      organization,
      title,
      bio,
      password,
      confirmPassword
    } = req.body || {};

    if (password && password !== confirmPassword) {
      return res.status(400).json({ message: '两次输入的密码不一致' });
    }

    const updates = [];
    const params = [];

    if (username !== undefined) {
      const sanitizedUsername = sanitizeField(username, MAX_FIELD_LENGTH.username);
      if (!sanitizedUsername) {
        return res.status(400).json({ message: '用户名不能为空' });
      }
      updates.push('username = ?');
      params.push(sanitizedUsername);
    }

    if (organization !== undefined) {
      updates.push('organization = ?');
      params.push(sanitizeField(organization, MAX_FIELD_LENGTH.organization));
    }

    if (title !== undefined) {
      updates.push('title = ?');
      params.push(sanitizeField(title, MAX_FIELD_LENGTH.title));
    }

    if (bio !== undefined) {
      updates.push('bio = ?');
      params.push(sanitizeField(bio, MAX_FIELD_LENGTH.bio));
    }

    if (password) {
      if (password.length < 8) {
        return res.status(400).json({ message: '密码长度至少为8位' });
      }
      const salt = await bcrypt.genSalt(10);
      const hashedPassword = await bcrypt.hash(password, salt);
      updates.push('password_hash = ?');
      params.push(hashedPassword);
    }

    if (!updates.length) {
      return res.status(400).json({ message: '没有检测到需要更新的字段' });
    }

    updates.push('updated_at = ?');
    params.push(new Date());
    params.push(userId);

    await mysqlPool.execute(
      `UPDATE users
       SET ${updates.join(', ')}
       WHERE id = ?`,
      params
    );

    if (req.user && updates.some(field => field.startsWith('username'))) {
      req.user.username = params[updates.findIndex(field => field.startsWith('username'))];
    }

    if (req.user && updates.some(field => field.startsWith('password_hash'))) {
      req.user.tokenVersion = (req.user.tokenVersion || 0) + 1;
    }

    const [users] = await mysqlPool.execute(
      'SELECT id, username, email, role, created_at, updated_at, organization, title, bio FROM users WHERE id = ?',
      [userId]
    );

    const updatedUser = users[0];

    res.status(200).json({
      message: '资料更新成功',
      user: {
        id: updatedUser.id,
        username: updatedUser.username,
        email: updatedUser.email,
        role: updatedUser.role,
        createdAt: updatedUser.created_at,
        updatedAt: updatedUser.updated_at,
        organization: updatedUser.organization,
        title: updatedUser.title,
        bio: updatedUser.bio
      }
    });
  } catch (error) {
    console.error('更新用户信息错误:', error);
    if (error.code === 'ER_DUP_ENTRY') {
      return res.status(409).json({ message: '用户名或邮箱已存在' });
    }
    res.status(500).json({ message: '服务器内部错误' });
  }
});

module.exports = router;